refactor: move logic out of SharedState in preparation for its removal
